test(providers,watch): add the same-package regression tests CI required
CI's require-test-for-fix gate was right: this PR changed source in
internal/providers and internal/watch with no test beside it, and a fix without a
regression test is a fix nobody can prove stays fixed.

internal/watch gets the two that matter for the daemon: a wedged notification
helper must not pause the watch loop, and a descendant it spawned must not
outlive the deadline. The descendant sleeps longer than the deadline on purpose —
an earlier draft had it finish first, so it proved nothing.

internal/providers gets the inverse assertion: browser launches must NOT be
waited on. A fake launcher that hangs for 30s has to return immediately, because
the window it opens is one a human is meant to keep open. The second test
confirms the failure that does matter — a missing launcher — still surfaces.

Also splits the Trainline helper-assisted fallback into its own file. Bounding it
pushed trainline.go to 805 lines, over the repo's 800-line limit; the fallback is
a self-contained path with its own budget and failure modes, so it is a real seam
rather than shuffling lines to satisfy a counter. 805 -> 681 plus 143.

Verified: full suite exit 0 across 99 packages; check-file-size.sh passes; vet,
staticcheck and golangci-lint clean; GOOS=windows vet clean.

Refs #507
